Enh: Tests - Change file installed, adapt old tests (http and ssl) and add new ones for http server

+++ test/test_sslv3_disabled.py
@@ -82,7 +82,6 @@ class testSchedulerInit(AlignakTest):
 
         d.load_config_file()
 
-        d.http_backend = 'cherrypy'
         d.do_daemon_init_and_start(fake=True)
         d.load_modules_manager()
 
@@ -91,20 +90,22 @@ class testSchedulerInit(AlignakTest):
         if not hasattr(ssl, 'SSLContext'):
             print 'BAD ssl version for testing, bailing out'
             return
-        ctx = ssl.SSLContext(ssl.PROTOCOL_SSLv3)
-        ctx.check_hostname=False
-        ctx.verify_mode=ssl.CERT_NONE
-        self.conn = httplib.HTTPSConnection("localhost:9998",context=ctx)
-        self.assertRaises(ssl.SSLError,self.conn.connect)
-        try:
-            self.conn.connect()
-        except ssl.SSLError as e:
-            assert e.reason == 'SSLV3_ALERT_HANDSHAKE_FAILURE'
-        sleep(2)
-        pid = int(file("tmp/arbiterd.pid").read())
-        print ("KILLING %d" % pid)*50
-        os.kill(int(file("tmp/arbiterd.pid").read()), 2)
-        d.do_stop()
+        # New version have PROTOCOL_SSLv23 that will open v2 by default so wont fail as we expect it
+        if hasattr(ssl, "PROTOCOL_SSLv3"):
+            ctx = ssl.SSLContext(ssl.PROTOCOL_SSLv3)
+            ctx.check_hostname = False
+            ctx.verify_mode = ssl.CERT_NONE
+            self.conn = httplib.HTTPSConnection("localhost:9998", context=ctx)
+            self.assertRaises(ssl.SSLError, self.conn.connect)
+            try:
+                self.conn.connect()
+            except ssl.SSLError as e:
+                assert e.reason == 'SSLV3_ALERT_HANDSHAKE_FAILURE'
+            sleep(2)
+            pid = int(file("tmp/arbiterd.pid").read())
+            print ("KILLING %d" % pid)*50
+            os.kill(int(file("tmp/arbiterd.pid").read()), 2)
+            d.do_stop()
 
 
 if __name__ == '__main__':
